requirements.txt
streamlit
tensorflow
ultralytics
roboflow
transformers
opencv-python
matplotlib
numpy
pandas
Pillow
pytesseract
